+++ R/utils_render_latex.R
@@ -447,23 +447,17 @@ create_columns_component_l <- function(data, colwidth_df) {
 
     if (length(stub_layout) > 1L) {
       # If stub_layout == 1L, multicolumn is not needed and stub_label is already defined
-      stub_df <- dplyr::filter(colwidth_df, type == 'stub')
-      stub_width <-
-        if (stub_df$pt > 0) {
-          sprintf("%.2fpt", stub_df$pt)
-        } else if (stub_df$lw > 0) {
-          sprintf("%.2f\\linewidth", stub_df$lw)
-        } else {
-          ""
-        }
-
-      if (stub_width == "") {
+      stub_df <- dplyr::filter(colwidth_df, type %in% c("stub", "row_group"))
+      if (any(stub_df$unspec == 1L)) {
         width_txt <- "c"
       } else {
         width_txt <-
           sprintf(
-            ">{\\centering\\arraybackslash}m{\\dimexpr %s -2\\tabcolsep-1.5\\arrayrulewidth}",
-            stub_width
+            ">{\\centering\\arraybackslash}m{%s}",
+            create_singlecolumn_width_text_l(
+              pt = sum(stub_df$pt),
+              lw = sum(stub_df$lw)
+            )
           )
       }
 
@@ -626,8 +620,9 @@ create_body_component_l <- function(data, colwidth_df) {
         n_rows_in_group <- n_rows_in_group + dim(list_of_summaries$summary_df_data_list[[i]])[1L]
       }
       row_splits_body[[groups_rows_df$row_start[i]]][1] <-
-        sprintf("\\multirow{%d}{=}{%s}",
+        sprintf("\\multirow{%d}{%s}{%s}",
                 n_rows_in_group,
+                if (colwidth_df$unspec[1L] == 1L) "*" else "=",
                 groups_rows_df$group_label[i])
         #groups_rows_df$group_label[i]
     }
@@ -1714,7 +1709,7 @@ create_colwidth_df_l <- function(data) {
     }
   }
 
-  if (length(stub_layout) > length(c('stub', 'row_group') %in% width_df$type)) {
+  if (length(stub_layout) > sum(c('stub', 'row_group') %in% width_df$type)) {
     if ('stub' %in% width_df$type) {
       stub_row_group <-
         dplyr::mutate(